+++ apps/app-frontend/src/components/ui/settings/display/AppearanceSettings.vue
