+++ apps/app-frontend/src/components/ui/settings/display/LanguageSettings.vue
