+++ apps/app-frontend/src/components/ui/settings/instances/DefaultInstanceSettings.vue
