+++ apps/app-frontend/src/components/ui/settings/instances/JavaSettings.vue
