+++ apps/app-frontend/src/components/ui/settings/instances/ResourceManagementSettings.vue
