+++ apps/app-frontend/src/locales/en-US/index.json
@@ -168,7 +168,7 @@
     "message": "Default landing page"
   },
   "app.appearance-settings.hide-nametag.description": {
-    "message": "Hide your username above the player preview on the Skins page."
+    "message": "Hide your username above the player preview on the Skin selector page."
   },
   "app.appearance-settings.hide-nametag.title": {
     "message": "Hide nametag"
@@ -842,51 +842,6 @@
   "app.settings.downloading": {
     "message": "Downloading v{version}"
   },
-  "app.settings.feature-flags.advanced-filters-collapsed": {
-    "message": "Collapse advanced filters"
-  },
-  "app.settings.feature-flags.always-show-app-controls": {
-    "message": "Always show app controls"
-  },
-  "app.settings.feature-flags.i18n-debug": {
-    "message": "Internationalization debug"
-  },
-  "app.settings.feature-flags.page-path": {
-    "message": "Page path"
-  },
-  "app.settings.feature-flags.pride-fundraiser": {
-    "message": "Pride fundraiser"
-  },
-  "app.settings.feature-flags.project-background": {
-    "message": "Project background"
-  },
-  "app.settings.feature-flags.reset-to-default": {
-    "message": "Reset to default"
-  },
-  "app.settings.feature-flags.server-project-qa": {
-    "message": "Server project QA"
-  },
-  "app.settings.feature-flags.server-ram-as-bytes-always-on": {
-    "message": "Always show server RAM in bytes"
-  },
-  "app.settings.feature-flags.show-instance-play-time": {
-    "message": "Show instance play time"
-  },
-  "app.settings.feature-flags.show-version-environment-column": {
-    "message": "Show version environment column"
-  },
-  "app.settings.feature-flags.skip-non-essential-warnings": {
-    "message": "Skip non-essential warnings"
-  },
-  "app.settings.feature-flags.skip-unknown-pack-warning": {
-    "message": "Skip unknown pack warning"
-  },
-  "app.settings.feature-flags.worlds-in-home": {
-    "message": "Worlds on Home"
-  },
-  "app.settings.feature-flags.worlds-tab": {
-    "message": "Worlds tab"
-  },
   "app.settings.java-installations.location.title": {
     "message": "Java {version, number} location"
   },
@@ -963,7 +918,7 @@
     "message": "Appearance"
   },
   "app.settings.tabs.behavior": {
-    "message": "Behaviour"
+    "message": "Behavior"
   },
   "app.settings.tabs.default-instance-options": {
     "message": "Default instance options"
